pyproject.toml
[project]
name = "auto-gen-explore"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "semantic-kernel (>=1.20.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "ollama (>=0.4.7,<0.5.0)",
    "autogen-agentchat (>=0.4.7,<0.5.0)",
    "autogen-core (>=0.4.7,<0.5.0)",
    "autogen-ext[docker,magentic-one,openai] (>=0.4.8.2,<0.5.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "nanoid (>=2.0.0,<3.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "azure-identity (>=1.20.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "yfinance (>=0.2.54,<0.3.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
